fixes alert component
