+++ src/components/pages/admins/index.js
@@ -2,12 +2,11 @@ import { useState, useEffect, } from "react";
 import { Form, } from "antd";
 import { getToken, } from "../../../util/auth";
 import { request, } from "../../../util/request";
-import { getMessage, } from "../../../util";
+import { emitMessage, } from "../../../util";
 import { sectionStyle, } from "../../../stitches.config";
 
 import Section from "../../core/Section";
 import Container from "../../core/Container";
-import Alert from "../../widgets/Alert";
 import Text from "../../core/Text";
 import RegisterAdmin from "../../widgets/RegisterAdmin";
 import UserCards from "../../widgets/UserCards";
@@ -29,17 +28,11 @@ export const Admins = ({ isAuth, authUser, }) => {
     const [isModalVisible, setIsModalVisible] = useState(false);
     const [modalContent, setModalContent] = useState('');
     const [title, setTitle] = useState('');
-    const [alert, setAlert] = useState('');
-    const [status, setStatus] = useState('');
-    const [header, setHeader] = useState('');
 
     const handleAdministrators = payload => setAdministrators(payload);
     const handleShowModal = () => setIsModalVisible(true);
     const handleHideModal = () => setIsModalVisible(false);
     const handleTitle = title => setTitle(title);
-    const handleAlert = message => setAlert(message);
-    const handleStatus = status => setStatus(status);
-    const handleHeader = header => setHeader(header);
 
     const handleModalContent = (payload, title) => {
         setModalContent(payload);
@@ -49,28 +42,6 @@ export const Admins = ({ isAuth, authUser, }) => {
 
     const resetForm = () => {
         form.resetFields();
-        handleHeader('');
-        handleStatus('');
-        handleAlert('');
-    }
-
-    const handleAlertComponent = (header, status, message) => {
-        if (!(message)) {
-            handleAlert('');
-            return;
-        }
-
-        handleHeader(header);
-        handleStatus(status);
-        handleAlert(<Text type="span">{message}</Text>);
-    }
-
-    const emitMessage = (content, status, duration) => {
-        return getMessage({
-            content: content,
-            status: status,
-            duration: duration,
-        });
     }
 
     useEffect(() => {
@@ -91,12 +62,6 @@ export const Admins = ({ isAuth, authUser, }) => {
     return (
         <Section css={sectionStyle}>
             <Container>
-            {
-                alert &&
-                <Alert status={status} header={header}>
-                    {alert}
-                </Alert>
-            }
                 <Button 
                 onClick={() => handleModalContent(
                     <RegisterAdmin
@@ -106,18 +71,9 @@ export const Admins = ({ isAuth, authUser, }) => {
                     emitMessage={emitMessage}
                     isAuth={isAuth}
                     resetForm={resetForm}
-                    handleAlertComponent={handleAlertComponent}
                     administrators={administrators}
                     handleAdministrators={handleAdministrators}
-                    handleHideModal={handleHideModal}
-                    {...alert && {
-                        alert: <Alert
-                            status={status}
-                            header={header}
-                            css={{ marginBottom: '$20', }}>
-                            {alert}
-                        </Alert>
-                    }} />, "Add Administrator"
+                    handleHideModal={handleHideModal} />, "Add Administrator"
                 )}
                 text="Add" />
             </Container>
@@ -130,7 +86,9 @@ export const Admins = ({ isAuth, authUser, }) => {
                 css={{ ...cardGroupStyling }}
                 onUpdate={toggleAdminStatus}
                 authUser={authUser}
-                emitMessage={emitMessage} /> :
+                emitMessage={emitMessage}
+                isModalVisible={isModalVisible}
+                handleModalContent={handleModalContent} /> :
                 <Text type="span">No data</Text>
             }
             </Container>
