+++ src/components/pages/students/index.js
@@ -21,17 +21,11 @@ export const Students = ({ isAuth, authUser, }) => {
     const [isModalVisible, setIsModalVisible] = useState(false);
     const [modalContent, setModalContent] = useState('');
     const [title, setTitle] = useState('');
-    const [alert, setAlert] = useState('');
-    const [status, setStatus] = useState('');
-    const [header, setHeader] = useState('');
 
     const handleStudents = payload => setStudents(payload);
     const handleShowModal = () => setIsModalVisible(true);
     const handleHideModal = () => setIsModalVisible(false);
     const handleTitle = title => setTitle(title);
-    const handleAlert = message => setAlert(message);
-    const handleStatus = status => setStatus(status);
-    const handleHeader = header => setHeader(header);
 
     const handleModalContent = (payload, title) => {
         setModalContent(payload);
@@ -41,20 +35,6 @@ export const Students = ({ isAuth, authUser, }) => {
 
     const resetForm = () => {
         form.resetFields();
-        handleHeader('');
-        handleStatus('');
-        handleAlert('');
-    }
-
-    const handleAlertComponent = (header, status, message) => {
-        if (!(message)) {
-            handleAlert('');
-            return;
-        }
-        
-        handleHeader(header);
-        handleStatus(status);
-        handleAlert(<Text type="span">{message}</Text>);
     }
 
     const emitMessage = (content, status, duration) => {
@@ -92,18 +72,9 @@ export const Students = ({ isAuth, authUser, }) => {
                     emitMessage={emitMessage}
                     isAuth={isAuth}
                     resetForm={resetForm}
-                    handleAlertComponent={handleAlertComponent}
                     students={students}
                     handleStudents={handleStudents}
-                    handleHideModal={handleHideModal}
-                    {...alert && {
-                        alert: <Alert
-                            status={status}
-                            header={header}
-                            css={{ marginBottom: '$20', }}>
-                            {alert}
-                        </Alert>
-                    }} />, "Add Student"
+                    handleHideModal={handleHideModal} />, "Add Student"
                 )}
                 text="Add" />
             </Container>
@@ -121,10 +92,6 @@ export const Students = ({ isAuth, authUser, }) => {
             maskClosable={false}
             title={title}
             onCancel={handleHideModal}>
-            {
-                alert &&
-                <Alert status={status} header={header}>{alert}</Alert>
-            }
                 {modalContent}
             </Modal>
         }
