+++ src/components/widgets/AdminSettings/index.js
@@ -1,28 +1,19 @@
 import { getToken, } from "../../../util/auth";
 import { request, } from "../../../util/request";
-import { getMessage, } from "../../../util";
 
 import SettingsNameForm from "../SettingsNameForm";
-import SettingsDisplayPhotoForm from "../SettingsDisplayPhotoForm";
 import SettingsEmailForm from "../SettingsEmailForm";
 import SettingsPasswordForm from "../SettingsPasswordForm";
 
-export const StudentSettings = ({ 
-    slug, 
+export const AdminSettings = ({
+    slug,
     authUser,
     values,
-    handleStudent,
+    handleAdministrator,
+    emitMessage,
 }) => {
-    const emitMessage = (content, status, duration) => {
-        return getMessage({
-            content: content,
-            status: status,
-            duration: duration,
-        });
-    }
-
     return (
-        (values && (Object.keys(values).length > 0)) && 
+        (values && (Object.keys(values).length > 0)) &&
         <>
             <SettingsNameForm
             authUser={authUser}
@@ -30,33 +21,25 @@ export const StudentSettings = ({
             onFinish={updateName}
             emitMessage={emitMessage}
             slug={slug}
-            handleArrayObj={handleStudent} />
-            <SettingsDisplayPhotoForm
-            authUser={authUser}
-            values={values}
-            onFinish={updateDisplayPhoto}
-            emitMessage={emitMessage}
-            slug={slug}
-            handleArrayObj={handleStudent} />
+            handleArrayObj={handleAdministrator} />
             <SettingsEmailForm
             authUser={authUser}
             values={values}
             onFinish={updateEmail}
             emitMessage={emitMessage}
             slug={slug}
-            handleArrayObj={handleStudent} />
+            handleArrayObj={handleAdministrator} />
             <SettingsPasswordForm
             authUser={authUser}
             onFinish={updatePassword}
             emitMessage={emitMessage}
             slug={slug} />
         </>
-
     )
 }
 
 async function updateName(form) {
-    return request.post("student-name-update", form, {
+    return request.post("admin-name-update", form, {
         headers: {
             'Authorization': `Bearer ${getToken()}`,
         }
@@ -64,7 +47,7 @@ async function updateName(form) {
 }
 
 async function updateEmail(form) {
-    return request.post("student-email-update", form, {
+    return request.post("admin-email-update", form, {
         headers: {
             'Authorization': `Bearer ${getToken()}`,
         }
@@ -72,19 +55,11 @@ async function updateEmail(form) {
 }
 
 async function updatePassword(form) {
-    return request.post("student-password-update", form, {
+    return request.post("admin-password-update", form, {
         headers: {
             'Authorization': `Bearer ${getToken()}`,
         }
     });
 }
 
-async function updateDisplayPhoto(form) {
-    return request.post("student-display-photo-update", form, {
-        headers: {
-            'Authorization': `Bearer ${getToken()}`,
-        }
-    });
-}
-
-export default StudentSettings;
+export default AdminSettings;
