+++ src/components/widgets/StudentPayments/index.js
@@ -31,11 +31,8 @@ export const StudentPayments = ({
     const handleHideModal = () => setIsModalVisible(false);
     const handleTitle = title => setTitle(title);
 
-    const resetForm = (form, handleHeader, handleStatus, handleAlert) => {
+    const resetForm = form => {
         form.resetFields();
-        handleHeader('');
-        handleStatus('');
-        handleAlert('');
     }
 
     const handleModalContent = (payload, title) => {
