+++ src/components/widgets/StudentPermitForm/index.js
@@ -10,7 +10,7 @@ import {
 } from "antd";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faCloudArrowUp, faCircleXmark, } from "@fortawesome/free-solid-svg-icons";
-import { getErrorMessage, } from "../../../util";
+import { getErrorMessage, getAlertComponent, } from "../../../util";
 import { styled, } from "../../../stitches.config";
 
 import Container from "../../core/Container";
@@ -67,14 +67,10 @@ export const StudentPermitForm = ({
     const [file, setFile] = useState('');
     const [imageUrl, setImageUrl] = useState('');
     const [alert, setAlert] = useState('');
-    const [status, setStatus] = useState('');
-    const [header, setHeader] = useState('');
 
     const handleFile = value => setFile(value);
     const handleImageUrl = value => setImageUrl(value);
-    const handleAlert = message => setAlert(message);
-    const handleStatus = status => setStatus(status);
-    const handleHeader = header => setHeader(header);
+    const handleAlertComponent = payload => setAlert(payload);
     let arr;
 
     const handleHelpers = payload => setHelpers(payload);
@@ -150,16 +146,16 @@ export const StudentPermitForm = ({
         },
     ];
 
-    const handleAlertComponent = (header, status, message) => {
-        if (!(message)) {
-            handleAlert('');
-            return;
-        }
+    // const handleAlertComponent = (header, status, message) => {
+    //     if (!(message)) {
+    //         handleAlert('');
+    //         return;
+    //     }
 
-        handleHeader(header);
-        handleStatus(status);
-        handleAlert(<Text type="span">{message}</Text>);
-    }
+    //     handleHeader(header);
+    //     handleStatus(status);
+    //     handleAlert(<Text type="span">{message}</Text>);
+    // }
 
     const handleRemoveImage = () => {
         handleFile('');
@@ -200,7 +196,7 @@ export const StudentPermitForm = ({
 
         onFinish(storeForm).then(response => {
             if (!(response.data.is_success)) {
-                handleAlertComponent("Error", "danger", response.data.data);
+                handleAlertComponent(getAlertComponent("Error", "danger", response.data.data, { marginTop: '0', }));
                 return;
             }
 
@@ -210,11 +206,11 @@ export const StudentPermitForm = ({
                 arr = [response.data.data.details]
             }
 
-            resetForm(form, handleHeader, handleStatus, handleAlert);
+            resetForm(form);
             handleRemoveImage();
             handlePermits(arr);
             handleHideModal();
-            handleAlertComponent("", "", null);
+            handleAlertComponent(getAlertComponent(null, null, null));
             setTimeout(() => {
                 emitMessage("Permit added.", "success", 2.5);
             }, 2000);
@@ -252,12 +248,9 @@ export const StudentPermitForm = ({
 
     return (
         <Container css={styling}>
-            <Container>
-            {
-                alert && 
-                <Alert status={status} header={header} css={{ margin: '0', }}>{alert}</Alert>
-            }
-            </Container>
+        {
+            alert
+        }
         {
             !(file) &&
             <Container>
