--- src/components/widgets/StudentRegistrarFile/index.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-export const StudentRegistrarFile = () => {
-    return (
-        <div>
-            File
-        </div>
-    )
-}
-
-export default StudentRegistrarFile;
