+++ src/util/index.js
@@ -9,6 +9,7 @@ import {
 import { message, } from "antd";
 import Container from "../components/core/Container";
 import Text from "../components/core/Text";
+import Alert from "../components/widgets/Alert";
 import { 
     fadeOut, 
     spinnerStyle,
@@ -123,4 +124,33 @@ export const paymentTypes = {
 export const statusOptions = {
     pending: "Pending",
     verified: "Verified",
+}
+
+export const getAlertComponent = (
+    header, 
+    status, 
+    message,
+    css,
+) => {
+    return ( 
+        <>
+        {
+            message && 
+            <Alert
+            status={status}
+            header={header}
+            css={{ margin: '$50 0px', ...css, }}>
+                {message}
+            </Alert>   
+        }
+        </>
+    )
+}
+
+export const emitMessage = (content, status, duration) => {
+    return getMessage({
+        content: content,
+        status: status,
+        duration: duration,
+    });
 }
